+++ CHANGES.md
@@ -12,6 +12,8 @@
 
 - Added option cfg.connRandomSecFromList=True to set if sec and loc are randomly chosen from list when synsPerConn=1 
 
+- Added cfg.printSynsAfterRule to print total connections after each conn rule is applied 
+
 - Fixed bug calculating avg pop rate in plotRaster when using subsets of pops
 
 - Fixed bug LFP recording only applied to compartmental cells